requirements.txt
python-telegram-bot==21.10
requests==2.32.5
python-dotenv==1.0.1
